Added accel & decel
When snappy move is disabled, the player will have a quick acceleration to max speed. When the player stops moving, they will slip just a little as they decelerate- if they instead hold in the opposite direction of where they're moving, they'll slip for a quick moment before accelerating in the opposite direction.